improve IMAP command conflict handling
From now on, IMAP command conflicts between different IMAP clients discovered
by `fetch` turn all following `delete`s into noops, instead of aborting
the whole program cycle, as it did before.

This way parallel IMAP sessions will no longer make `imaparms fetch` to
leave unmarked messages on the server and leftovers in `./tmp` of `--maildir`.

The only other consistent alternative solution to this is to delete messages
in `./tmp` instead, but then `imaparms` will have to re-download them again
in the next cycle, which is wasteful.

Also, that previous "just abort it" behaviour could get really annoying when
conflicts between clients are common, i.e. if `--every` intervals of IMAP
clients intersect a lot, which is easy to achieve when they all start at system
boot and the intervals are short and/or close to having a common divisor.