*, vendor/kisstdlib: start using `kisstdlib`, simplify, improve logging
A bunch of code from here was imported to `kisstdlib` and improved
there. So, switching to using that as improves log message output and error
handling quite a bit. Though, it requires `CPython >= 3.11` now.

Also, `kisstdlib` added support for per-subcommand `--help`, so that
also works now, which is nice.

This also makes some lines previously written to `stderr` to be written
to `stdout` instead, adds more log messages, and improves some of the old ones.
All of this is done in this one commit because it's too inter-dependent to
split into smaller pieces.